Refactor Card styles, fix bug with Navbar notification, move icons to new folders
