+++ src/components/Card/Card.module.scss
@@ -84,7 +84,7 @@
     color: $primary;
   }
 
-  &__fullPrice {
+  &__full_price {
     font-size: 22px;
     font-family: 'Mont-Regular', sans-serif;
     font-weight: 600;
