+++ src/components/Header/Navbar/Navbar.module.scss
@@ -81,22 +81,22 @@
     width: 300px;
     margin-right: 20px;
     flex-grow: 1;
+  }
 
-    &__notification_badge {
-      position: absolute;
-      top: 25%;
-      left: 50%;
-      width: 14px;
-      height: 14px;
-      background-color: red;
-      color: white;
-      border-radius: 50%;
-      display: flex;
-      border: 1px solid white;
-      align-items: center;
-      justify-content: center;
-      font-size: 9px;
-    }
+  &__notification_badge {
+    position: absolute;
+    top: 25%;
+    left: 50%;
+    width: 14px;
+    height: 14px;
+    background-color: red;
+    color: white;
+    border-radius: 50%;
+    display: flex;
+    border: 1px solid white;
+    align-items: center;
+    justify-content: center;
+    font-size: 9px;
   }
   
   &__hidden {
